Move GeometryConverter into schema pkg
